% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netdose-package.R
\docType{package}
\name{netdose-package}
\alias{netdose-package}
\title{netdose: Brief overview of methods}
\description{
R package \bold{netdose} provides methods and graphical tools to conduct
the network meta-analysis with dose-response relationships in a frequentist way.
}
\details{
R package \bold{netdose} is a tool to conduct dose-response network meta-analysis
a frequentist way (Petropoulou et al, 2025). The package can implement the dose-response
network meta-analysis model (function \code{\link{netdose}}); calculate the predicted values of the dose-response
network meta-analysis model (function \code{\link{predict}}); provide dose-response plots
(function \code{\link{plot.netdose}}) (Petropoulou et al., 2025).

Type \code{help(package = "netdose")} for a listing of R functions
available in \bold{netdose}.

Type \code{citation("netdose")} on how to cite \bold{NMAoutlier}
in publications.

To report problems and bugs, please send an email to Dr. Maria
Petropoulou <maria.petropoulou@uniklinik-freiburg.de>.

The development version of \bold{netdose} is available on GitHub
\url{https://github.com/petropouloumaria/netdose}.
}
\references{
Petropoulou et al. (2025):
Network meta-analysis with dose-response relationships.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petropouloumaria/netdose}
}

}
\author{
Petropoulou Maria <maria.petropoulou@.uniklinik-freiburg.de>,
  Guido Schwarzer <guido.schwarzer@uniklinik-freiburg.de>
}
\keyword{package}
