% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predict.netdose.R
\name{print.predict.netdose}
\alias{print.predict.netdose}
\title{Print method for objects of class predict}
\usage{
\method{print}{predict.netdose}(
  x,
  backtransf = attr(x, "backtransf"),
  digits = gs("digits"),
  big.mark = gs("big.mark"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{predict.netdose}.}

\item{backtransf}{A logical indicating whether printed results
should be back transformed. If \code{backtransf=TRUE}, results
for \code{sm="OR"} are printed as odds ratios rather than log
odds ratios, for example.}

\item{digits}{Minimal number of significant digits, see
\code{print.default}.}

\item{big.mark}{A character used as thousands separator.}

\item{\dots}{Additional arguments (ignored).}
}
\description{
Print method for objects of class predict
}
