% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Dot plot for dose-response data}
\usage{
dotplot(
  x,
  col = "black",
  fill = "orange",
  size = 2,
  drop.reference.group = TRUE,
  ylab = "Dose",
  ...
)
}
\arguments{
\item{x}{An object created with \code{\link{netdose}}.}

\item{col}{The color used for the border of dots.}

\item{fill}{The color used for the background of dots.}

\item{size}{A single numeric with the size of the dots.}

\item{drop.reference.group}{A logical indicating whether to drop the
panel for the reference group.}

\item{ylab}{A label for the y-axis.}

\item{\dots}{Additional arguments (ignored).}
}
\value{
No return value.
}
\description{
Generates a dot plot for dose-response data
}
\details{
The function produces a dot plot of drug doses.
}
\note{
The message \emph{'Bin width defaults to 1/30 of the range of the data.
  Pick better value with `binwidth`.'} is irrelevant as dot sizes should
  be identical. Setting the argument 'binwidth' would result in different
  dot sizes for the drugs.
}
\examples{
# Use a subset of 5 studies from anesthesia data
anesthesia_subset <-
  subset(anesthesia, study \%in\% unique(anesthesia$study)[1:5])

# Prepare data for DR-NMA
dat <- pairwise(
  agent = list(agent1, agent2, agent3),
  event = list(event1, event2, event3),
  n = list(n1, n2, n3),
  dose = list(dose1, dose2, dose3),
  data = anesthesia_subset,
  studlab = study,
  append = FALSE)

# DR-NMA with linear dose-response function
dr1 <- netdose(TE, seTE, agent1, dose1, agent2,
  dose2, studlab,
  data = dat)

# Dose-response plot
dotplot(dr1)

}
\seealso{
\code{\link{netdose}}
}
\author{
Guido Schwarzer <guido.schwarzer@uniklinik-freiburg.de>
}
\keyword{hplot}
